pyproject.toml
[project]
name = "ragops-exam"
version = "0.1.0"
description = "Solution for the ragops course exam"
readme = "README.md"
requires-python = ">=3.8.10"
dependencies = [
    "requests>=2.31.0",
    "fpdf>=1.7.2",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "psycopg2-binary>=2.9.9",
    "minio>=7.2.0",
    "anyio==4.4.0",
    "httpx==0.27.0",
    "datasets>=2.18.0",
    "ragas>=0.1.9",
    "langchain-openai>=0.1.0",
]